Randomization + UI Re-Design
This branch commit made over 1000+ lines of code change to the randomization logic to make it much easier to modify in the future. This is not a 100% implementation, but it is a move in the right direction. This also updates the UI to make it more streamlined for the end user.